refactor: separate app into multiple files
Separating the `app` object creation, `db` object creation, and route creation will make it easier to import the parts we need when we start putting things into more files. As-is, it's super easy to accidentally create import cycles that are annoying and difficult to break.
