+++ arlo_server/base.py
@@ -1,0 +1,4 @@
+from flask import Flask
+from config import STATIC_FOLDER
+
+app = Flask(__name__, static_folder=STATIC_FOLDER)
